+++ releases.md
@@ -2,7 +2,8 @@
 
 ## Features
 
-- Added new syntax for literal big ints and big decimals: `100000N`, `100000.30M` 
+- Added new syntax for literal big ints and big decimals: `100000N`, `100000.30M`
+- Added support for big ints and big decimals in pattern matching
 
 # 0.2.0